+++ src/main/java/com/hust/ewsystem/service/impl/ModelsServiceImpl.java
@@ -129,7 +129,7 @@ public class ModelsServiceImpl extends ServiceImpl<ModelsMapper, Models> impleme
                         .setStartTime(startTimeDate)
                         .setEndTime(endTimeDate);
                 tasksMapper.insert(newtask);
-                Integer taskId= newtask.getTaskId();
+                Long taskId= newtask.getTaskId();
                 File taskDir = new File(pythonFilePath + "/task_logs/" + taskLabel);
                 if (!taskDir.exists()) {
                     if (!taskDir.mkdirs()) {
@@ -430,7 +430,7 @@ public class ModelsServiceImpl extends ServiceImpl<ModelsMapper, Models> impleme
                     .setTaskLabel(taskLabel)
                     .setStartTime(LocalDateTime.now());
             tasksMapper.insert(newtask);
-            Integer taskId= newtask.getTaskId();
+            Long taskId= newtask.getTaskId();
             File taskDir = new File(pythonFilePath + "/task_logs/" + taskLabel);
             if (!taskDir.exists()) {
                 if (!taskDir.mkdirs()) {
@@ -502,7 +502,7 @@ public class ModelsServiceImpl extends ServiceImpl<ModelsMapper, Models> impleme
             e.printStackTrace();
         }
     }
-    public void executePredict(String filepath, String algorithmLabel, String taskLabel,Integer modelId,Integer taskId) {
+    public void executePredict(String filepath, String algorithmLabel, String taskLabel,Integer modelId,Long taskId) {
         Process process = null;
         boolean interrupted = false;
         try {
@@ -558,7 +558,7 @@ public class ModelsServiceImpl extends ServiceImpl<ModelsMapper, Models> impleme
             }
         }
     }
-    private void readAndSaveResults(String filepath, String taskLabel,Integer modelId,Integer taskId) {
+    private void readAndSaveResults(String filepath, String taskLabel,Integer modelId,Long taskId) {
         try {
             String resultFilePath = filepath + "/task_logs/" + taskLabel + "/result.json";
 
@@ -621,7 +621,7 @@ public class ModelsServiceImpl extends ServiceImpl<ModelsMapper, Models> impleme
      * @param taskId taskId
      * @param formatter formatter
      */
-    private void processAlerts(List<JSONObject> alertList, int modelId, int taskId,DateTimeFormatter formatter) {
+    private void processAlerts(List<JSONObject> alertList, int modelId, Long taskId,DateTimeFormatter formatter) {
         Iterator<JSONObject> iterator = alertList.iterator();
         while (iterator.hasNext()) {
             JSONObject alert = iterator.next();
